Split emit direct validation tests
